No call hierarchy for method reference after fixing compile error #438
SourceIndexer.indexResolvedDocument() adds an index entry for a method
reference only if the expression is resolved. If the expression becomes
resolved later on, e.g. due to fixing compile errors in another file,
SourceIndexer is not triggered again. So the index for the file remains
incomplete and the method reference is not found e.g. when computing the
call hierarchy of the referenced method.

This change adds a new index category, an unresolved method reference.
References of methods with the same name and different arguments count
will not be distinguished in this new category, as the argument count is
not known at the time of adding the unresolved method reference to the
index.

Fixes: #438

Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>