Wrong call hierarchy computed for type in conflicting modules
JavaSearchNameEnvironment adds module infos to each PackageFragmentRoot
of a project, whether the root fragment is on the compile classpath or
module path. However the module info is only valid for fragments that
are on the module path.

As a result, a search (call hierarchy, references) can encounter
non-existent compile errors such as module conflicts. The project itself
compiles fine, but the search detects such conflicts due to the module
infos attached to each PackageFragmentRoot.

This change adjusts JavaSearchNameEnvironment to not add module info for
a PackageFragmentRoot which is not on the module path of the project.

Fixes: #675
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>