Wrong search result for module conflict in classpath referenced projects
WIP, DON'T MERGE

This change fixes wrong search results when referenced projects on the
project classpath have a module conflict. It does so by providing the
project for which the search was initiated to JavaSearchNameEnvironment,
so that the search can know which referenced projects are on the
classpath and which are on the module path.

Fixes: #675